Consolidate status testing (#3048)
* Consolidate {ASSERT,EXPECT_{,NOT_}_OK macros
