+++ Firestore/core/test/firebase/firestore/remote/serializer_test.cc
@@ -45,6 +45,7 @@
 #include "Firestore/core/src/firebase/firestore/util/status.h"
 #include "Firestore/core/src/firebase/firestore/util/statusor.h"
 #include "Firestore/core/test/firebase/firestore/testutil/testutil.h"
+#include "Firestore/core/test/firebase/firestore/util/status_testing.h"
 #include "absl/types/optional.h"
 #include "google/protobuf/stubs/common.h"
 #include "google/protobuf/util/message_differencer.h"
@@ -73,11 +74,6 @@ using firebase::firestore::testutil::Key;
 using firebase::firestore::util::Status;
 using firebase::firestore::util::StatusOr;
 using google::protobuf::util::MessageDifferencer;
-
-#define ASSERT_OK(status) ASSERT_TRUE(StatusOk(status))
-#define ASSERT_NOT_OK(status) ASSERT_FALSE(StatusOk(status))
-#define EXPECT_OK(status) EXPECT_TRUE(StatusOk(status))
-#define EXPECT_NOT_OK(status) EXPECT_FALSE(StatusOk(status))
 
 TEST(Serializer, CanLinkToNanopb) {
   // This test doesn't actually do anything interesting as far as actually using
@@ -114,30 +110,6 @@ class SerializerTest : public ::testing::Test {
       const SnapshotVersion& read_time,
       const v1::BatchGetDocumentsResponse& proto) {
     ExpectDeserializationRoundTrip(key, absl::nullopt, read_time, proto);
-  }
-
-  /**
-   * Checks the status. Don't use directly; use one of the relevant macros
-   * instead. eg:
-   *
-   *   Status good_status = ...;
-   *   ASSERT_OK(good_status);
-   *
-   *   Status bad_status = ...;
-   *   EXPECT_NOT_OK(bad_status);
-   */
-  testing::AssertionResult StatusOk(const Status& status) {
-    if (!status.ok()) {
-      return testing::AssertionFailure()
-             << "Status should have been ok, but instead contained "
-             << status.ToString();
-    }
-    return testing::AssertionSuccess();
-  }
-
-  template <typename T>
-  testing::AssertionResult StatusOk(const StatusOr<T>& status) {
-    return StatusOk(status.status());
   }
 
   /**
