+++ Firestore/core/test/firebase/firestore/util/filesystem_test.cc
@@ -25,7 +25,7 @@
 #include "Firestore/core/src/firebase/firestore/util/log.h"
 #include "Firestore/core/src/firebase/firestore/util/path.h"
 #include "Firestore/core/src/firebase/firestore/util/string_win.h"
-#include "Firestore/core/test/firebase/firestore/util/status_test_util.h"
+#include "Firestore/core/test/firebase/firestore/util/status_testing.h"
 #include "absl/strings/match.h"
 #include "absl/types/optional.h"
 #include "gtest/gtest.h"
