+++ Firestore/core/test/firebase/firestore/util/status_apple_test.mm
@@ -19,7 +19,7 @@
 #include <cerrno>
 
 #include "Firestore/core/src/firebase/firestore/util/error_apple.h"
-#include "Firestore/core/test/firebase/firestore/util/status_test_util.h"
+#include "Firestore/core/test/firebase/firestore/util/status_testing.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
